Fixes unsignedBigIntefer foreign key error

+++ database/migrations/2019_05_07_133946_create_beers_table.php
@@ -23,10 +23,10 @@ class CreateBeersTable extends Migration
             $table->float('plato')->nullable();
             $table->integer('stock')->default(0);
 
-            $table->unsignedInteger('color_id')->nullable();
-            $table->unsignedInteger('brewery_id')->nullable();
-            $table->unsignedInteger('packaging_id')->nullable();
-            $table->unsignedInteger('style_id')->nullable();
+            $table->unsignedBigInteger('color_id')->nullable();
+            $table->unsignedBigInteger('brewery_id')->nullable();
+            $table->unsignedBigInteger('packaging_id')->nullable();
+            $table->unsignedBigInteger('style_id')->nullable();
 
             $table->timestamps();
